cli: allow adding workspace in existing empty dir
Cloning into an existing empty directory works. There is no reason why
'jj workspace add' should fail. After all, the directory is empty.

For example, if the target directory is the root of a file system,
it is impossible to remove directory to let 'jj workspace add' create
it. Reword the error message to match the one for jj git clone.