split: warn instead of failing on empty commit with path argument
If the user does something like `jj split @` and the working copy is
not empty, they will get warnings like this:

```
Warning: No matching entries for paths: @
Warning: No changes have been selected, so the new revision will be empty
```

If they run the same command when the working copy is empty, they will
instead get this:

```
Error: Refusing to split empty commit e518dba7418a1b20a545c12ecb16cc870756efa3.
Hint: Use `jj new` if you want to create another empty commit.
```

That's probably less helpful because it's not clear that the `@`
argument was interpreted as a path.

The rationale for the error (in 07559f24ece9) is that it's confusing
that an empty diff editor pops up if you accidentally try to split an
empty commit. This patch changes it so we print a warning when the
diff is empty instead of starting the diff editor.