Adjust misleading `gix-credentials` shell script fixture shebangs
These shell scripts are all written for `sh` and they are currently
always invoked as arguments to an explicit `sh` interpreter (either
`sh`, or a dynamically determined `sh.exe` path on Windows). But
they had a `bash` shebang.

This changes their shebang to the simple shebang for `sh` to avoid
confusion. This is not expected to change the behavior of the test
suite; it is mainly for clarity.

(This change shouldn't be made to most other test fixture scripts;
they're intended for `bash`, often using `bash`-specific features.)