Don't run `shell_builtin_or_command_in_path*` on Windows
This makes the `shell_builtin_or_command_in_path*` tests only run
on Unix-like systems. Although they're able to pass in some
commonly used Windows environments, including Git Bash and the
environment used in GitHub Actions runners, they actually rely on
`printf` (and, for the recently added zero-argument one, `echo`)
being present in a `PATH` directory.

This should always be the case on Unix-like systems (at least if
attempting to abide by the POSIX requirement that such commands be
available externally whether or nor they are also shell builtins),
but it is often not the case on Windows, and when it isn't, this
doesn't indicate any problem with the Windows environment.

Variants of these tests have recently been added that ensure a
shell is actually used: both implicitly, by prepending `:;` to the
commands so manual argument splitting cannot be used instead of
running a shell, and (in other test cases) explicitly, by calling
`with_shell()`, which suppreses argument splitting.

Therefore, sufficient coverage for the functionality being tested
here is present on Windows even without the versions that take the
form of calling a builtin but actually in practice rely on an
external command.

This turns those off on Windows, but keeps them on Unix-like
systems since they may guard against regressions if an attempt is
ever made to automatically detect efforts to call a shell builtin
and suppress implicit argument splitting. (However, this is not to
say that such a feature ought to be implemented, nor to predict
that it will: most likely this would not make sense to add, since
it would be hard to get right, but also beacuse it would diverge
from the behavior of Git. But if Git ever does this, then the
regression tests against doing it incorrectly may be helpful.)