Use quoted absolute path for `cat` on Windows in filter tests
Two test cases use `cat` as a clean and smudge filter, to check for
regressions of #2080 (fixed in #2262). These tests passed on
Unix-like systems. They also passed in Windows environments where a
`cat` command is in `PATH`, including Git Bash environments and the
full environments on Windows CI runners. However, they failed with
"program not found" errors in Windows environments where `cat` is
not in a `PATH` directory.

The purpose of the two tests cases is not to make sure the exact
command `cat` can be used successfully as a clean and smudge filter
on all systems, but instead to produce a condition that caused a
deadlock (#2080) prior to the improvements in #2262. Therefore, it
can be considered a test bug that the tests were often unable to
specify a usable `cat` command on Windows (even when one existed).

This commit fixes that test bug, by resolving `cat` to a full path
that uses `/` separators but that is otherwise a Windows-style
path, and by single quoting the path for use in a shell, passing
that instead of the literal command `cat`. If it cannot be found,
`cat.exe` is used as a fallback. Non-Windows systems continue
always to use `cat`.

The following tests that previously passed on Windows when `cat` is
in `PATH`, but failed on Windows when `cat` is not in `PATH`, now
pass on Windows whether `cat` is in `PATH` or not:

    gix-filter::filter driver::apply::large_file_with_cat_filter_does_not_hang
    gix-filter::filter driver::apply::large_file_with_cat_filter_early_drop

The technique used to resolve `cat` to the a usable full path is
somewhat ad-hoc. At least for now, it runs a command in a shell
that uses the shell to look up `cat` and when uses `cygpath` to
resolve it. (`cygpath` supports making Windows paths with `/`
separators with the `--mixed`/`-m` option.)